refactor(sales): decompose SalesService into Query/Creation/Writer + DiscountPolicy util
Behavior-preserving money/$tx. Facade keeps 8-method surface + (prisma,interCompany)
ctor + delegates; internally constructs sub-services (forwardRef + module + spec untouched).
3 per-type writer $tx moved WHOLE with EXACT isolation (cash/external Serializable,
installment DEFAULT); interCompanyService.createFromSaleInTx cross-seam tx threaded;
loyalty post-commit best-effort $tx + _loyaltyRedemptionFailed flag preserved;
costPrice-strip/OWNER-profit role shaping intact. tsc 0, sales tests green, lint 0.
